% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\name{adj}
\alias{adj}
\title{Generates Adjacency Matrix}
\usage{
adj(A, y, n, mfc)
}
\arguments{
\item{A}{Load the technical-coefficient matrix}

\item{y}{Load the finaldemand}

\item{n}{Numbers of rounds}

\item{mfc}{Minimal flow condition or the cut-off arbitarily choosen to binarize the layer matrix}
}
\value{
The function provides the adjacency matrix in which various graph-theories designs can be implemented.
}
\description{
It generates the adjancency matrix based on the layer matrix and binarization on some arbitarily set up cut-off value.
The layer matrix having value more than such arbitary cut-off is set to 1 and 0 otherwise.
}

