% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbt_tr.R
\name{cbt_tr}
\alias{cbt_tr}
\title{Total Requirement Matrices - Commodity Based Tecnhology}
\usage{
cbt_tr(make, use)
}
\arguments{
\item{make}{Load the make matrix}

\item{use}{Load the use matrix}
}
\value{
Generates lists of various commodity based technology total requirement matrices.
}
\description{
Based on commodity based technology assumptions,this function generates various total requirement matrices on commodities-by-commodities, commodities-by-industries, industries-by-commodities and industries-by-industries.
}

