% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag2vec.R
\name{diag2vec}
\alias{diag2vec}
\title{Extract Diagonal of a Matrix}
\usage{
diag2vec(x)
}
\arguments{
\item{x}{an input matrix}
}
\value{
Extracts the diagonal of a matrix.
}
\description{
Given an input matrix, diag2vec returns a column vector of the elements along the diagonal.
}

