% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getB.R
\name{getB}
\alias{getB}
\title{Commodity by Industry Direct Requirement}
\usage{
getB(make, use)
}
\arguments{
\item{make}{Load the make matrix}

\item{use}{Load the use matrix}
}
\value{
Generates commodity by industry direct requirement matrix. Use Latex formula later.
}
\description{
Generates commodity by industry direct requirement matrix
}

