% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getC.R
\name{getC}
\alias{getC}
\title{Industries Output Proportions Matrix}
\usage{
getC(make)
}
\arguments{
\item{make}{Load the make matrix}
}
\value{
Generates industries output proportions matrix also known as product mix or commodity mix matrix. Use Latex formula later. The column sum of this matrix thus should be unity. It has commodities-by-industries dimension.
}
\description{
Generates industries output proportions matrix also known as product mix or commodity mix matrix.
}

