% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getD.R
\name{getD}
\alias{getD}
\title{Commodities Output Proportions Matrix}
\usage{
getD(make)
}
\arguments{
\item{make}{Load the make matrix}
}
\value{
Generates industries output proportions matrix also known as product mix or commodity mix matrix. Use Latex formula later.It has industries-by-commodities dimension.
}
\description{
Generates commodities output proportions matrix also known as market shares matrix.
}

