% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibt_tr.R
\name{ibt_tr}
\alias{ibt_tr}
\title{Total Requirement Matrices - Industry Based Tecnhology}
\usage{
ibt_tr(make, use)
}
\arguments{
\item{make}{Load the make matrix}

\item{use}{Load the use matrix}
}
\value{
Generates lists of various industry based technology total requirement matrices.
}
\description{
Based on industry based technology assumptions,this function generates various total requirement matrices on commodities-by-commodities, commodities-by-industries, industries-by-commodities and industries-by-industries.
}

