% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impcoef.R
\name{impcoef}
\alias{impcoef}
\title{Generates Importance Coefficient}
\usage{
impcoef(IminusAInv, y, supress = 0)
}
\arguments{
\item{IminusAInv}{Load total requirement matrix.}

\item{y}{Load the output vector}

\item{supress}{An arbitary cutoff}
}
\value{
The function returns the importance coefficient bases on the total requirement matrix. The algorithm follows Aroche-Reyes, 1996; Schintke & Staglin, 1988. Thus tracks the importance between i-th sector and j-sector.
}
\description{
It generates the importance coefficient bases on the total requirement matrix.
}

