% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOcheckdim.R
\name{IOcheckdim}
\alias{IOcheckdim}
\title{Check the Consistency of Dimensions of Input Matrix}
\usage{
IOcheckdim(make, use, finaldemand, valueadded)
}
\arguments{
\item{make}{Load the make matrix}

\item{use}{Load the use matrix}

\item{finaldemand}{Load the finaldemand matrix}

\item{valueadded}{Load the valueadded matrix}
}
\value{
Checks if the numbers of rows of use and finaldemand matrix are same or not.

Checks if the numbers of row of use  and numbers of columns of make make matrix are same or not.

Checks if the numbers of columns of use and numbers of columns of valueadded matrix are same or not.
}
\description{
Check the consistency of the dimension of Make, Use, Final Demand and Valueadded Matrix.
}

