% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOchecknum.R
\name{IOchecknum}
\alias{IOchecknum}
\title{Check for existence of Non-numerics is Input Matrix}
\usage{
IOchecknum(make, use, finaldemand, valueadded)
}
\arguments{
\item{make}{Load the make matrix}

\item{use}{Load the use matrix}

\item{finaldemand}{Load the finaldemand matrix}

\item{valueadded}{Load the valueadded matrix}
}
\value{
Check the existence of Non-numerics in Make, Use, Final Demand and Valueadded Matrix.
}
\description{
Check the existence of Non-numerics in Make, Use, Final Demand and Valueadded Matrix.
}

