% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOchecksum.R
\name{IOchecksum}
\alias{IOchecksum}
\title{Check for balances of Input Matrix}
\usage{
IOchecksum(make, use, finaldemand, valueadded)
}
\arguments{
\item{make}{Load the make matrix}

\item{use}{Load the use matrix}

\item{finaldemand}{Load the finaldemand matrix}

\item{valueadded}{Load the valueadded matrix}
}
\value{
Checks if the sums of "row sums of use matrix and row sums of final demand" is same with "column sum of make matrix".

Checks if the sums of "column sums of use matrix and column sums of valueadded matrix" is same with "row sums of make matrix".
}
\description{
Check the balances of total commodities inputs and outputs and the balances of industry inputs and outputs.
}

