% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOfullcheck.R
\name{IOfullcheck}
\alias{IOfullcheck}
\title{Check the feasibilities of Input Matricies}
\usage{
IOfullcheck(make, use, finaldemand, valueadded)
}
\arguments{
\item{make}{Load the make matrix}

\item{use}{Load the use matrix}

\item{finaldemand}{Load the finaldemand matrix}

\item{valueadded}{Load the valueadded matrix}
}
\description{
Check the consistency of the dimension, existence on numerical values and sum balances of Make, Use, Final Demand and Valueadded Matrix.
}

