% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer}
\alias{layer}
\title{Generates Layer Matrix}
\usage{
layer(A, y, n, mfc)
}
\arguments{
\item{A}{Load the technical-coefficient matrix}

\item{y}{Load the finaldemand}

\item{n}{Numbers of rounds}

\item{mfc}{Minimal Flow Condition}
}
\value{
The layer matrix reflects the intermediary flow triggered by the final demand in some N-th round of the process. Therefore
they are in absolute volume amonunts and interpretable as delivery flows and comparable to each other.
}
\description{
The layer matrix reflects the intermediary flow triggered by the final demand at some N-th round of the process.
}

