% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matequal.R
\name{matequal}
\alias{matequal}
\title{Check If Two Matrix Are Same or Not.}
\usage{
matequal(x, y)
}
\arguments{
\item{x}{is a matrix}

\item{y}{is another matrix to be tested if it's same with x or not.}
}
\value{
TRUE if matrix are identical and FALSE otherwise.
}
\description{
Considers two matrix x and y and checks if the matrix are identical in term of dimension and elements or not.
}

