% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlayer.R
\name{rlayer}
\alias{rlayer}
\title{Generates Layer Matrice Recursively}
\usage{
rlayer(A, y, k, mfc)
}
\arguments{
\item{A}{Load the technical-coefficient matrix}

\item{y}{Load the finaldemand}

\item{k}{Numbers of rounds to iterate}

\item{mfc}{Minimla Flow Control}
}
\value{
This function returns all the rounds of layer matrices and adjacency matrices based on the given minimal flow control parameter. The layer matrix reflects the intermediary flow triggered by the final demand in some N-th round of the process. Therefore
they are in absolute volume amonunts and interpretable as delivery flows and comparable to each other.
}
\description{
The layer matrix reflects the intermediary flow triggered by the final demand at some N-th round of the process.
}

