% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/techcoeff.R
\name{techcoeff}
\alias{techcoeff}
\title{Generates Technical Coefficient}
\usage{
techcoeff(IminusAInv)
}
\arguments{
\item{IminusAInv}{Load total requirement matrix.}
}
\value{
The function returns the technical coefficient bases on the total requirement matrix. Note: Make sure to use correct inverse of (I-A).
}
\description{
It generates the technical coefficient bases on the total requirement matrix.
}

