% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2diag.R
\name{vec2diag}
\alias{vec2diag}
\title{Create Diagonal Matrix From Vector}
\usage{
vec2diag(x)
}
\arguments{
\item{x}{a row or a column vector}
}
\value{
Elemnts of x are placed along the diagonal.
}
\description{
Given an input row or column vector, vec2diag returns a diagonal matrix with the input argument along the diagonal.
}

